Move the crash point conflict check to its own file
The check depends on neither the crash point nor logging, so inside the scenario
class it ran eight times under names that did not apply to it. A sibling class in
the same file avoided that but left the file with two test classes, which only 3%
of the suite does.
